requirements.txt
Flask==3.0.0
Pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
WTForms==3.1.1
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2
webdriver-manager==4.0.1
markdown==3.5.1
weasyprint==60.2
reportlab==4.0.7 